#44: implement Klarna transactions Unit tests
