+++ payone-adapter/src/test/java/com/commercetools/payment/payone/methods/transaction/PayoneKlarnaCreatePaymentTransactionMethodProviderTest.java
@@ -1,0 +1,21 @@
+package com.commercetools.payment.payone.methods.transaction;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
+
+@RunWith(MockitoJUnitRunner.class)
+public class PayoneKlarnaCreatePaymentTransactionMethodProviderTest
+        extends BasePayoneCreatePaymentTransactionMethodProviderTest {
+
+    @Before
+    public void setUp() throws Exception {
+        transactionMethod = PayoneKlarnaCreatePaymentTransactionMethodProvider.of();
+    }
+
+    @Test
+    public void handleSuccessfulServiceCall() throws Exception {
+        handleSuccessfulServiceCall_default();
+    }
+}
